pyproject.toml
[project]
name = "nicetomeetyou"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.3.3",
    "redis>=6.4.0",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "nba_news.settings_test"
python_files = ["test_*.py", "*_tests.py"]
testpaths = ["news/tests"]
